guidance-for-codex-on-amazon-bedrock/source/pyproject.toml
[tool.poetry]
name = "codex-with-bedrock"
version = "1.0.0"
description = "Deploy and manage OpenAI Codex CLI on Amazon Bedrock"
authors = ["Codex Team <codex@example.com>"]
readme = "../README.md"
license = "MIT"
keywords = ["aws", "bedrock", "codex", "openai", "ai", "authentication", "monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [
    { include = "credential_provider" },
    { include = "codex_with_bedrock" }
]

[tool.poetry.scripts]
codex-with-bedrock = "codex_with_bedrock.cli:main"
cxwb = "codex_with_bedrock.cli:main"
credential-provider = "credential_provider:main"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
boto3 = "^1.42.88"
requests = "^2.33.1"
PyJWT = "^2.8.0"
keyring = "^24.0.0"
cryptography = ">=44.0.1"
cleo = "^2.0.0"
rich = "^14.3.3"
pydantic = "^2.12.5"
pyyaml = "^6.0"
questionary = "^2.1.1"
cfn-flip = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.3"
pytest-cov = "^7.1.0"
black = "^26.3.1"
nuitka = "^4.0.8"
ordered-set = "^4.1.0"
zstandard = "^0.25.0"
pyinstaller = "^6.19.0"
ruff = "^0.15.10"
mypy = "^1.20.0"
types-requests = "*"
pre-commit = "^4.5.1"
cfn-lint = "^1.48.1"
yamllint = "^1.38.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.black]
line-length = 120
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true